services/bridge/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async support
anyio>=4.2.0
httpx>=0.26.0

# SSE streaming
sse-starlette>=1.8.0

# Configuration
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0

# Optional: for token counting
tiktoken>=0.5.0
python-multipart
